requirements.txt
pycryptodome
PyQt5